Moved CLI Builder classes to CLI Builder project
